+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 3.3.3 ##
 
-1.  UI Toolkit now can fall back to an IMGUI custom drawer if the target field is specified to be drawn by an IMGUI drawer (but label in `RichLabel` might not work as it's depended on the IMGUI drawer itself, especially `<color>` & `<icon>`)
+1.  UI Toolkit now can fall back to an IMGUI custom drawer if the target field is specified to be drawn by an IMGUI drawer (note: `<icon>` in `RichLabel` will not work and will get removed)
 
 ## 3.3.2 ##